rename class and prepare backward compatibility
